requirements.txt
numpy==1.13.3
Pillow==4.3.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
